refactor logging code and counter int -> long

+++ app/src/main/java/com/dokechin/tasktimer/com/dokechin/tasktimer/domain/Counter.java
@@ -4,15 +4,15 @@ package com.dokechin.tasktimer.com.dokechin.tasktimer.domain;
  * Created by dokechin on 2016/07/22.
  */
 public class Counter {
-    private int sec = 0;
-    public Counter(int second) {
+    private long sec = 0;
+    public Counter(long second) {
         this.sec = second;
     }
     public String clockFormat(){
 
-        int hour = this.sec / 3600;
-        int min = (this.sec  - hour * 3600) / 60;
-        int sec = (this.sec  - hour * 3600 - min * 60);
+        long hour = this.sec / 3600;
+        long min = (this.sec  - hour * 3600) / 60;
+        long sec = (this.sec  - hour * 3600 - min * 60);
 
         return String.format("%1$02d:%2$02d:%3$02d", hour, min , sec);
     }
